requirements.txt
fastapi==0.129.2
uvicorn[standard]==0.41.0
# Use google-generativeai only
google-generativeai>=0.8.6
python-dotenv
